+++ sdk/formrecognizer/Azure.AI.FormRecognizer/tests/FormRecognizerClient/FormRecognizerClientLiveTests.cs
@@ -155,7 +155,7 @@ namespace Azure.AI.FormRecognizer.Tests
                 Assert.GreaterOrEqual(cell.Confidence, 0, $"Cell with text {cell.Text} should have confidence greater or equal to zero.");
                 Assert.LessOrEqual(cell.RowIndex, 1, $"Cell with text {cell.Text} should have a row index less than or equal to one.");
 
-                Assert.Greater(cell.TextContent.Count, 0, $"Cell with text {cell.Text} should have text content.");
+                Assert.Greater(cell.FieldElements.Count, 0, $"Cell with text {cell.Text} should have at least one field element.");
             }
         }
 
@@ -245,7 +245,7 @@ namespace Azure.AI.FormRecognizer.Tests
                 Assert.IsFalse(cell.IsHeader, $"Cell with text {cell.Text} should not have been classified as header.");
 
                 Assert.GreaterOrEqual(cell.Confidence, 0, $"Cell with text {cell.Text} should have confidence greater or equal to zero.");
-                Assert.Greater(cell.TextContent.Count, 0, $"Cell with text {cell.Text} should have text content.");
+                Assert.Greater(cell.FieldElements.Count, 0, $"Cell with text {cell.Text} should have at least one field element.");
             }
         }
 
@@ -279,7 +279,7 @@ namespace Azure.AI.FormRecognizer.Tests
             {
                 var formPage = formPages[pageIndex];
 
-                ValidateFormPage(formPage, includeTextContent: true, expectedPageNumber: pageIndex + 1);
+                ValidateFormPage(formPage, includeFieldElements: true, expectedPageNumber: pageIndex + 1);
 
                 // Basic sanity test to make sure pages are ordered correctly.
 
@@ -294,7 +294,7 @@ namespace Azure.AI.FormRecognizer.Tests
         public async Task StartRecognizeContentCanParseBlankPage()
         {
             var client = CreateFormRecognizerClient();
-            var options = new RecognizeOptions() { IncludeTextContent = true };
+            var options = new RecognizeOptions() { IncludeFieldElements = true };
             RecognizeContentOperation operation;
 
             using var stream = FormRecognizerTestEnvironment.CreateStream(TestFile.Blank);
@@ -306,7 +306,7 @@ namespace Azure.AI.FormRecognizer.Tests
             FormPageCollection formPages = await operation.WaitForCompletionAsync(PollingInterval);
             var blankPage = formPages.Single();
 
-            ValidateFormPage(blankPage, includeTextContent: true, expectedPageNumber: 1);
+            ValidateFormPage(blankPage, includeFieldElements: true, expectedPageNumber: 1);
 
             Assert.AreEqual(0, blankPage.Lines.Count);
             Assert.AreEqual(0, blankPage.Tables.Count);
@@ -332,7 +332,7 @@ namespace Azure.AI.FormRecognizer.Tests
             {
                 var formPage = formPages[pageIndex];
 
-                ValidateFormPage(formPage, includeTextContent: true, expectedPageNumber: pageIndex + 1);
+                ValidateFormPage(formPage, includeFieldElements: true, expectedPageNumber: pageIndex + 1);
 
                 // Basic sanity test to make sure pages are ordered correctly.
 
@@ -435,7 +435,7 @@ namespace Azure.AI.FormRecognizer.Tests
             Assert.AreEqual("Itemized", form.Fields["ReceiptType"].Value.AsString());
             Assert.AreEqual("Contoso Contoso", form.Fields["MerchantName"].Value.AsString());
             Assert.AreEqual("123 Main Street Redmond, WA 98052", form.Fields["MerchantAddress"].Value.AsString());
-            Assert.AreEqual("123-456-7890", form.Fields["MerchantPhoneNumber"].ValueText.Text);
+            Assert.AreEqual("123-456-7890", form.Fields["MerchantPhoneNumber"].ValueData.Text);
 
             var date = form.Fields["TransactionDate"].Value.AsDate();
             var time = form.Fields["TransactionTime"].Value.AsTime();
@@ -542,7 +542,7 @@ namespace Azure.AI.FormRecognizer.Tests
             Assert.AreEqual("Itemized", form.Fields["ReceiptType"].Value.AsString());
             Assert.AreEqual("Contoso Contoso", form.Fields["MerchantName"].Value.AsString());
             Assert.AreEqual("123 Main Street Redmond, WA 98052", form.Fields["MerchantAddress"].Value.AsString());
-            Assert.AreEqual("987-654-3210", form.Fields["MerchantPhoneNumber"].ValueText.Text);
+            Assert.AreEqual("987-654-3210", form.Fields["MerchantPhoneNumber"].ValueData.Text);
 
             var date = form.Fields["TransactionDate"].Value.AsDate();
             var time = form.Fields["TransactionTime"].Value.AsTime();
@@ -601,7 +601,7 @@ namespace Azure.AI.FormRecognizer.Tests
         public async Task StartRecognizeReceiptsCanParseMultipageForm(bool useStream)
         {
             var client = CreateFormRecognizerClient();
-            var options = new RecognizeOptions() { IncludeTextContent = true };
+            var options = new RecognizeOptions() { IncludeFieldElements = true };
             RecognizeReceiptsOperation operation;
 
             if (useStream)
@@ -629,7 +629,7 @@ namespace Azure.AI.FormRecognizer.Tests
 
                 Assert.NotNull(recognizedForm);
 
-                ValidateRecognizedForm(recognizedForm, includeTextContent: true,
+                ValidateRecognizedForm(recognizedForm, includeFieldElements: true,
                     expectedFirstPageNumber: expectedPageNumber, expectedLastPageNumber: expectedPageNumber);
 
                 // Basic sanity test to make sure pages are ordered correctly.
@@ -638,8 +638,8 @@ namespace Azure.AI.FormRecognizer.Tests
                 {
                     var sampleField = recognizedForm.Fields["MerchantAddress"];
 
-                    Assert.IsNotNull(sampleField.ValueText);
-                    Assert.AreEqual("2345 Dogwood Lane Birch, Kansas 98123", sampleField.ValueText.Text);
+                    Assert.IsNotNull(sampleField.ValueData);
+                    Assert.AreEqual("2345 Dogwood Lane Birch, Kansas 98123", sampleField.ValueData.Text);
                 }
                 else if (formIndex == 1)
                 {
@@ -652,7 +652,7 @@ namespace Azure.AI.FormRecognizer.Tests
         public async Task StartRecognizeReceiptsCanParseBlankPage()
         {
             var client = CreateFormRecognizerClient();
-            var options = new RecognizeOptions() { IncludeTextContent = true };
+            var options = new RecognizeOptions() { IncludeFieldElements = true };
             RecognizeReceiptsOperation operation;
 
             using var stream = FormRecognizerTestEnvironment.CreateStream(TestFile.Blank);
@@ -665,7 +665,7 @@ namespace Azure.AI.FormRecognizer.Tests
 
             var blankForm = recognizedForms.Single();
 
-            ValidateRecognizedForm(blankForm, includeTextContent: true,
+            ValidateRecognizedForm(blankForm, includeFieldElements: true,
                 expectedFirstPageNumber: 1, expectedLastPageNumber: 1);
 
             Assert.AreEqual(0, blankForm.Fields.Count);
@@ -680,7 +680,7 @@ namespace Azure.AI.FormRecognizer.Tests
         public async Task StartRecognizeReceiptsCanParseMultipageFormWithBlankPage()
         {
             var client = CreateFormRecognizerClient();
-            var options = new RecognizeOptions() { IncludeTextContent = true };
+            var options = new RecognizeOptions() { IncludeFieldElements = true };
             RecognizeReceiptsOperation operation;
 
             using var stream = FormRecognizerTestEnvironment.CreateStream(TestFile.InvoiceMultipageBlank);
@@ -700,7 +700,7 @@ namespace Azure.AI.FormRecognizer.Tests
 
                 Assert.NotNull(recognizedForm);
 
-                ValidateRecognizedForm(recognizedForm, includeTextContent: true,
+                ValidateRecognizedForm(recognizedForm, includeFieldElements: true,
                     expectedFirstPageNumber: expectedPageNumber, expectedLastPageNumber: expectedPageNumber);
 
                 // Basic sanity test to make sure pages are ordered correctly.
@@ -708,10 +708,10 @@ namespace Azure.AI.FormRecognizer.Tests
                 if (formIndex == 0 || formIndex == 2)
                 {
                     var sampleField = recognizedForm.Fields["MerchantName"];
-                    var expectedValueText = formIndex == 0 ? "Bilbo Baggins" : "Frodo Baggins";
+                    var expectedValueData = formIndex == 0 ? "Bilbo Baggins" : "Frodo Baggins";
 
-                    Assert.IsNotNull(sampleField.ValueText);
-                    Assert.AreEqual(expectedValueText, sampleField.ValueText.Text);
+                    Assert.IsNotNull(sampleField.ValueData);
+                    Assert.AreEqual(expectedValueData, sampleField.ValueData.Text);
                 }
             }
 
@@ -760,10 +760,10 @@ namespace Azure.AI.FormRecognizer.Tests
         [TestCase(true, false)]
         [TestCase(false, true)]
         [TestCase(false, false)]
-        public async Task StartRecognizeCustomFormsWithLabels(bool useStream, bool includeTextContent)
+        public async Task StartRecognizeCustomFormsWithLabels(bool useStream, bool includeFieldElements)
         {
             var client = CreateFormRecognizerClient();
-            var options = new RecognizeOptions { IncludeTextContent = includeTextContent };
+            var options = new RecognizeOptions { IncludeFieldElements = includeFieldElements };
             RecognizeCustomFormsOperation operation;
 
             await using var trainedModel = await CreateDisposableTrainedModelAsync(useTrainingLabels: true);
@@ -789,7 +789,7 @@ namespace Azure.AI.FormRecognizer.Tests
 
             RecognizedForm form = operation.Value.Single();
 
-            ValidateRecognizedForm(form, includeTextContent: includeTextContent,
+            ValidateRecognizedForm(form, includeFieldElements: includeFieldElements,
                 expectedFirstPageNumber: 1, expectedLastPageNumber: 1);
 
             // Testing that we shuffle things around correctly so checking only once per property.
@@ -805,7 +805,7 @@ namespace Azure.AI.FormRecognizer.Tests
             var name = "PurchaseOrderNumber";
             Assert.IsNotNull(form.Fields[name]);
             Assert.AreEqual(FieldValueType.String, form.Fields[name].Value.Type);
-            Assert.AreEqual("948284", form.Fields[name].ValueText.Text);
+            Assert.AreEqual("948284", form.Fields[name].ValueData.Text);
         }
 
         [Test]
@@ -814,7 +814,7 @@ namespace Azure.AI.FormRecognizer.Tests
         public async Task StartRecognizeCustomFormsWithLabelsCanParseMultipageForm(bool useStream)
         {
             var client = CreateFormRecognizerClient();
-            var options = new RecognizeOptions() { IncludeTextContent = true };
+            var options = new RecognizeOptions() { IncludeFieldElements = true };
             RecognizeCustomFormsOperation operation;
 
             await using var trainedModel = await CreateDisposableTrainedModelAsync(useTrainingLabels: true, useMultipageFiles: true);
@@ -837,7 +837,7 @@ namespace Azure.AI.FormRecognizer.Tests
 
             var recognizedForm = recognizedForms.Single();
 
-            ValidateRecognizedForm(recognizedForm, includeTextContent: true,
+            ValidateRecognizedForm(recognizedForm, includeFieldElements: true,
                 expectedFirstPageNumber: 1, expectedLastPageNumber: 2);
 
             // Check some values to make sure that fields from both pages are being populated.
@@ -866,7 +866,7 @@ namespace Azure.AI.FormRecognizer.Tests
         public async Task StartRecognizeCustomFormsWithLabelsCanParseBlankPage()
         {
             var client = CreateFormRecognizerClient();
-            var options = new RecognizeOptions() { IncludeTextContent = true };
+            var options = new RecognizeOptions() { IncludeFieldElements = true };
             RecognizeCustomFormsOperation operation;
 
             await using var trainedModel = await CreateDisposableTrainedModelAsync(useTrainingLabels: true);
@@ -881,7 +881,7 @@ namespace Azure.AI.FormRecognizer.Tests
 
             var recognizedForm = recognizedForms.Single();
 
-            ValidateRecognizedForm(recognizedForm, includeTextContent: true,
+            ValidateRecognizedForm(recognizedForm, includeFieldElements: true,
                 expectedFirstPageNumber: 1, expectedLastPageNumber: 1);
 
             var blankPage = recognizedForm.Pages.Single();
@@ -896,7 +896,7 @@ namespace Azure.AI.FormRecognizer.Tests
         public async Task StartRecognizeCustomFormsWithLabelsCanParseMultipageFormWithBlankPage(bool useStream)
         {
             var client = CreateFormRecognizerClient();
-            var options = new RecognizeOptions() { IncludeTextContent = true };
+            var options = new RecognizeOptions() { IncludeFieldElements = true };
             RecognizeCustomFormsOperation operation;
 
             await using var trainedModel = await CreateDisposableTrainedModelAsync(useTrainingLabels: true);
@@ -919,7 +919,7 @@ namespace Azure.AI.FormRecognizer.Tests
 
             var recognizedForm = recognizedForms.Single();
 
-            ValidateRecognizedForm(recognizedForm, includeTextContent: true,
+            ValidateRecognizedForm(recognizedForm, includeFieldElements: true,
                 expectedFirstPageNumber: 1, expectedLastPageNumber: 3);
 
             for (int pageIndex = 0; pageIndex < recognizedForm.Pages.Count; pageIndex++)
@@ -976,10 +976,10 @@ namespace Azure.AI.FormRecognizer.Tests
         [TestCase(true, false)]
         [TestCase(false, true)]
         [TestCase(false, false)]
-        public async Task StartRecognizeCustomFormsWithoutLabels(bool useStream, bool includeTextContent)
+        public async Task StartRecognizeCustomFormsWithoutLabels(bool useStream, bool includeFieldElements)
         {
             var client = CreateFormRecognizerClient();
-            var options = new RecognizeOptions { IncludeTextContent = includeTextContent };
+            var options = new RecognizeOptions { IncludeFieldElements = includeFieldElements };
             RecognizeCustomFormsOperation operation;
 
             await using var trainedModel = await CreateDisposableTrainedModelAsync(useTrainingLabels: false);
@@ -1005,7 +1005,7 @@ namespace Azure.AI.FormRecognizer.Tests
 
             RecognizedForm form = operation.Value.Single();
 
-            ValidateRecognizedForm(form, includeTextContent: includeTextContent,
+            ValidateRecognizedForm(form, includeFieldElements: includeFieldElements,
                 expectedFirstPageNumber: 1, expectedLastPageNumber: 1);
 
             //testing that we shuffle things around correctly so checking only once per property
@@ -1020,9 +1020,9 @@ namespace Azure.AI.FormRecognizer.Tests
             Assert.IsNotNull(form.Fields);
             var name = "field-0";
             Assert.IsNotNull(form.Fields[name]);
-            Assert.IsNotNull(form.Fields[name].LabelText.Text);
+            Assert.IsNotNull(form.Fields[name].LabelData.Text);
             Assert.AreEqual(FieldValueType.String, form.Fields[name].Value.Type);
-            Assert.AreEqual("Hero Limited", form.Fields[name].LabelText.Text);
+            Assert.AreEqual("Hero Limited", form.Fields[name].LabelData.Text);
         }
 
         [Test]
@@ -1031,7 +1031,7 @@ namespace Azure.AI.FormRecognizer.Tests
         public async Task StartRecognizeCustomFormsWithoutLabelsCanParseMultipageForm(bool useStream)
         {
             var client = CreateFormRecognizerClient();
-            var options = new RecognizeOptions() { IncludeTextContent = true };
+            var options = new RecognizeOptions() { IncludeFieldElements = true };
             RecognizeCustomFormsOperation operation;
 
             await using var trainedModel = await CreateDisposableTrainedModelAsync(useTrainingLabels: false, useMultipageFiles: true);
@@ -1059,19 +1059,19 @@ namespace Azure.AI.FormRecognizer.Tests
                 var recognizedForm = recognizedForms[formIndex];
                 var expectedPageNumber = formIndex + 1;
 
-                ValidateRecognizedForm(recognizedForm, includeTextContent: true,
+                ValidateRecognizedForm(recognizedForm, includeFieldElements: true,
                     expectedFirstPageNumber: expectedPageNumber, expectedLastPageNumber: expectedPageNumber);
 
                 // Basic sanity test to make sure pages are ordered correctly.
 
                 var sampleField = recognizedForm.Fields["field-2"];
-                var expectedLabelText = formIndex == 0 ? "__Tokens__1" : "Contact:";
-                var expectedValueText = formIndex == 0 ? "Vendor Registration" : "Jamie@southridgevideo.com";
+                var expectedLabelData = formIndex == 0 ? "__Tokens__1" : "Contact:";
+                var expectedValueData = formIndex == 0 ? "Vendor Registration" : "Jamie@southridgevideo.com";
 
-                Assert.IsNotNull(sampleField.LabelText);
-                Assert.AreEqual(expectedLabelText, sampleField.LabelText.Text);
-                Assert.IsNotNull(sampleField.ValueText);
-                Assert.AreEqual(expectedValueText, sampleField.ValueText.Text);
+                Assert.IsNotNull(sampleField.LabelData);
+                Assert.AreEqual(expectedLabelData, sampleField.LabelData.Text);
+                Assert.IsNotNull(sampleField.ValueData);
+                Assert.AreEqual(expectedValueData, sampleField.ValueData.Text);
             }
         }
 
@@ -1079,7 +1079,7 @@ namespace Azure.AI.FormRecognizer.Tests
         public async Task StartRecognizeCustomFormsWithoutLabelsCanParseBlankPage()
         {
             var client = CreateFormRecognizerClient();
-            var options = new RecognizeOptions() { IncludeTextContent = true };
+            var options = new RecognizeOptions() { IncludeFieldElements = true };
             RecognizeCustomFormsOperation operation;
 
             await using var trainedModel = await CreateDisposableTrainedModelAsync(useTrainingLabels: false);
@@ -1094,7 +1094,7 @@ namespace Azure.AI.FormRecognizer.Tests
 
             var blankForm = recognizedForms.Single();
 
-            ValidateRecognizedForm(blankForm, includeTextContent: true,
+            ValidateRecognizedForm(blankForm, includeFieldElements: true,
                 expectedFirstPageNumber: 1, expectedLastPageNumber: 1);
 
             Assert.AreEqual(0, blankForm.Fields.Count);
@@ -1111,7 +1111,7 @@ namespace Azure.AI.FormRecognizer.Tests
         public async Task StartRecognizeCustomFormsWithoutLabelsCanParseMultipageFormWithBlankPage(bool useStream)
         {
             var client = CreateFormRecognizerClient();
-            var options = new RecognizeOptions() { IncludeTextContent = true };
+            var options = new RecognizeOptions() { IncludeFieldElements = true };
             RecognizeCustomFormsOperation operation;
 
             await using var trainedModel = await CreateDisposableTrainedModelAsync(useTrainingLabels: false);
@@ -1139,7 +1139,7 @@ namespace Azure.AI.FormRecognizer.Tests
                 var recognizedForm = recognizedForms[formIndex];
                 var expectedPageNumber = formIndex + 1;
 
-                ValidateRecognizedForm(recognizedForm, includeTextContent: true,
+                ValidateRecognizedForm(recognizedForm, includeFieldElements: true,
                     expectedFirstPageNumber: expectedPageNumber, expectedLastPageNumber: expectedPageNumber);
 
                 // Basic sanity test to make sure pages are ordered correctly.
@@ -1147,12 +1147,12 @@ namespace Azure.AI.FormRecognizer.Tests
                 if (formIndex == 0 || formIndex == 2)
                 {
                     var sampleField = recognizedForm.Fields["field-0"];
-                    var expectedValueText = formIndex == 0 ? "300.00" : "3000.00";
+                    var expectedValueData = formIndex == 0 ? "300.00" : "3000.00";
 
-                    Assert.IsNotNull(sampleField.LabelText);
-                    Assert.AreEqual("Subtotal:", sampleField.LabelText.Text);
-                    Assert.IsNotNull(sampleField.ValueText);
-                    Assert.AreEqual(expectedValueText, sampleField.ValueText.Text);
+                    Assert.IsNotNull(sampleField.LabelData);
+                    Assert.AreEqual("Subtotal:", sampleField.LabelData.Text);
+                    Assert.IsNotNull(sampleField.ValueData);
+                    Assert.AreEqual(expectedValueData, sampleField.ValueData.Text);
                 }
             }
 
@@ -1220,7 +1220,7 @@ namespace Azure.AI.FormRecognizer.Tests
             Assert.Throws<RequestFailedException>(() => operation.Value.GetType());
         }
 
-        private void ValidateFormPage(FormPage formPage, bool includeTextContent, int expectedPageNumber)
+        private void ValidateFormPage(FormPage formPage, bool includeFieldElements, int expectedPageNumber)
         {
             Assert.AreEqual(expectedPageNumber, formPage.PageNumber);
 
@@ -1232,7 +1232,7 @@ namespace Azure.AI.FormRecognizer.Tests
 
             Assert.NotNull(formPage.Lines);
 
-            if (!includeTextContent)
+            if (!includeFieldElements)
             {
                 Assert.AreEqual(0, formPage.Lines.Count);
             }
@@ -1284,14 +1284,14 @@ namespace Azure.AI.FormRecognizer.Tests
                     Assert.That(cell.Confidence, Is.LessThanOrEqualTo(1.0).Within(0.01));
 
                     Assert.NotNull(cell.Text);
-                    Assert.NotNull(cell.TextContent);
+                    Assert.NotNull(cell.FieldElements);
 
-                    if (!includeTextContent)
+                    if (!includeFieldElements)
                     {
-                        Assert.AreEqual(0, cell.TextContent.Count);
+                        Assert.AreEqual(0, cell.FieldElements.Count);
                     }
 
-                    foreach (var content in cell.TextContent)
+                    foreach (var content in cell.FieldElements)
                     {
                         Assert.AreEqual(expectedPageNumber, content.PageNumber);
                         Assert.NotNull(content.BoundingBox.Points);
@@ -1304,7 +1304,7 @@ namespace Azure.AI.FormRecognizer.Tests
             }
         }
 
-        private void ValidateRecognizedForm(RecognizedForm recognizedForm, bool includeTextContent, int expectedFirstPageNumber, int expectedLastPageNumber)
+        private void ValidateRecognizedForm(RecognizedForm recognizedForm, bool includeFieldElements, int expectedFirstPageNumber, int expectedLastPageNumber)
         {
             Assert.NotNull(recognizedForm.FormType);
             Assert.AreEqual(expectedFirstPageNumber, recognizedForm.PageRange.FirstPageNumber);
@@ -1318,7 +1318,7 @@ namespace Azure.AI.FormRecognizer.Tests
             for (int pageIndex = 0; pageIndex < recognizedForm.Pages.Count; pageIndex++)
             {
                 var formPage = recognizedForm.Pages[pageIndex];
-                ValidateFormPage(formPage, includeTextContent, expectedPageNumber);
+                ValidateFormPage(formPage, includeFieldElements, expectedPageNumber);
 
                 expectedPageNumber++;
             }
@@ -1337,39 +1337,39 @@ namespace Azure.AI.FormRecognizer.Tests
                 Assert.That(field.Confidence, Is.GreaterThanOrEqualTo(0.0).Within(0.01));
                 Assert.That(field.Confidence, Is.LessThanOrEqualTo(1.0).Within(0.01));
 
-                var labelText = field.LabelText;
+                var labelData = field.LabelData;
 
-                if (labelText != null)
+                if (labelData != null)
                 {
-                    Assert.Greater(labelText.PageNumber, 0);
+                    Assert.Greater(labelData.PageNumber, 0);
 
-                    if (labelText.BoundingBox.Points != null)
+                    if (labelData.BoundingBox.Points != null)
                     {
-                        Assert.AreEqual(4, labelText.BoundingBox.Points.Length);
+                        Assert.AreEqual(4, labelData.BoundingBox.Points.Length);
                     }
 
-                    Assert.NotNull(labelText.TextContent);
+                    Assert.NotNull(labelData.FieldElements);
 
-                    if (!includeTextContent)
+                    if (!includeFieldElements)
                     {
-                        Assert.AreEqual(0, labelText.TextContent.Count);
+                        Assert.AreEqual(0, labelData.FieldElements.Count);
                     }
                 }
 
-                var valueText = field.ValueText;
+                var valueData = field.ValueData;
 
-                Assert.NotNull(valueText);
+                Assert.NotNull(valueData);
 
-                if (valueText.BoundingBox.Points != null)
+                if (valueData.BoundingBox.Points != null)
                 {
-                    Assert.AreEqual(4, valueText.BoundingBox.Points.Length);
+                    Assert.AreEqual(4, valueData.BoundingBox.Points.Length);
                 }
 
-                Assert.NotNull(valueText.TextContent);
+                Assert.NotNull(valueData.FieldElements);
 
-                if (!includeTextContent)
+                if (!includeFieldElements)
                 {
-                    Assert.AreEqual(0, valueText.TextContent.Count);
+                    Assert.AreEqual(0, valueData.FieldElements.Count);
                 }
             }
         }