+++ sdk/formrecognizer/Azure.AI.FormRecognizer/CHANGELOG.md
@@ -6,9 +6,15 @@
 
 - Property `RequestedOn` renamed to `TrainingStartedOn` on `CustomFormModel` and `CustomFormModelInfo`.
 - Property `CompletedOn` renamed to `TrainingCompletedOn` on `CustomFormModel` and `CustomFormModelInfo`.
+- Property `LabelText` renamed to `LabelData` on `FormField`.
+- Property `ValueText` renamed to `ValueData` on `FormField`.
+- Property `TextContent` renamed to `FieldElements` on `FieldData` and `FormTableCell`.
 - Parameter `formUrl` in `StartRecognizeContent` has been renamed to `formUri`.
 - Parameter `receiptUrl` in `StartRecognizeReceipts` has been renamed to `receiptUri`.
 - Parameter `accessToken` in `CopyAuthorization.FromJson` has been renamed to `copyAuthorization`.
+- Parameter `IncludeTextContent` in `RecognizeOptions` has been renamed to `IncludeFieldElements`.
+- Model `FieldText` renamed to `FieldData`.
+- Model `FormContent` renamed to `FormElement`.
 
 ### Other breaking changes